Add InferenceExecRequest.__repr__() for debug logging (#958)
In the cycle leading up to the 3.2.0 release, we had some trouble
triaging CI error reports / routing bugs to the appropriate teams /
people. This and #917 are part of that.

The plan is to make shortfin's llm server log all relevant
InferenceExecRequests when something (e.g. kvcache allocations / nans in
logits) doesn't check out.